Add ability to define a callback on mail_server_protocols hook
